Refactor sportify services
